LED's Finally Working Again
So it seemed that the white 5mm LED I got from Ronson which he got from RigidWare was a fraud. Either I somehow killed it the first time i played with it (which is highly possible)  or the batch is fraudulent. Anyways, by using some substitute LED's from my dad's stash of thingies I re-did the workshop activity and managed to turn on the light.
I learned the following things: voltage supply is a given by the port you are using. For example, GPIO is rated for 3.3v and so is the power supply from STM32. Current is a result of the voltage and the total circuit resistance, coming from Ohm's law V=IR. Each resistor will drop the voltage, and putting them in series or parallel will either increase or decrease the combined resistance, where 1/R = 1/R1 + 1/R2 + ... 1/Rn. This is important because certain devices have specific ratings. Voltage and current can be adjusted to fit these devices, but if the device is out of the scope, meaning its required voltage is more than that can be supplied, then it can't be used. For diodes like LEDs, the brightness depends on how much current it receives, and how much current it receives has an exponential relation to the voltage difference across the device. These graphs can be found in their datasheets. This means that lower current = dimmer, higher current = brighter, but current must be above the minimum rating for it to shine at all, and below maximum rating to prevent the device from exploding. With the low electricity we are working with, we are given a lot of freedom on the number of resistors we can use, as long as our V=IR results stay within scope.

Since the 5mm LED was not working, I could at least confirm the LED bar was.