LED Bar Testing, More Pins Added
ELECTRICAL CHANGES:
The LED bar runs the same as any LED. It can use 120 Ohm resistors or even 66 Ohms for brighter (there is a range in which these products can operate in). However, the bar contains 10 LEDS, and controlling each one means I would need 20 cables and 10 resistors, which is a lot! I tried thinking of ways to control 10 LEDs with only 1 input, but that would require raising the incoming voltage to more than 3.3, since I would connect each LED in series. The reasoning is that each LED in the bar drops the voltage by a set constant (assuming), so the higher the input voltage the more LEDs turn on. However, the max output of the STM likely cannot support 10 LEDs in series, so that is not quite the option.
Instead, I brute forced and connected each LED on the bar with an input wire coming from GPIO output (so also 10 output pins). However, the other side of the circuit connected to ground (after electricity passes LED bar) is all connected in parallel. This means that LED #1 shares the same end circuit as LED #2 and #3... etc., and they run through the same resistor to the ground. This works since each LED drops the voltage voltage by the same constant amount. However, this parallelism does decrease the brightness of LEDs as more turn on, but not by much. Thus, I've effectively reduced 20 wires and 10 resistors down to 20 wires but 3 resistors, each 120 Ohms.

Since limit switch is always true by default (pressed down = open circuit), I tried using a Pull-Up resistor in the input, but that didn't quite work. Pull Up resistors can only exist near the input source, being 3.3v, but our input pins are actually ground, which can only use Pull Down resistors. In the end, I just negated the boolean value of the limit switch.

PROGRAM CHANGES:
Two functions were added, spamClick() and ledBar(int n). ledBar takes in an integer which represents the number of LED's to turn on. For example, n = 8 means 8 LED's in the bar, in order, will be turned on.
spamClick() is a function that continuously depletes/increases the LED bar. When GPIO PA12 (supposedly the limit switch) is true, the bar will increase, like a stack. When GPIO PA12 is false, the bar will decrease.
counter() has also been modified. The count number now only peaks at 10 and minimizes at 0. Increasing the counter increases the LED bar respectively, to the boundaries of the interval of counter.
password(int pin) has also been updated to show remaining tries on the LED bar.